FinalRound-AI/backend/schemas.py
from pydantic import BaseModel
from typing import Optional

class UserCreate(BaseModel):
    email: str
    password: str

class User(BaseModel):
    id: int
    email: str

    class Config:
        orm_mode = True

class Token(BaseModel):
    access_token: str
    token_type: str

class TokenData(BaseModel):
    email: Optional[str] = None

class InterviewRequest(BaseModel):
    job_role: str

class AnswerRequest(BaseModel):
    question: str
    answer: str
