FinalRound-AI/backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary
SQLAlchemy
pydantic-settings